IDE: fixed loading non-project files passed on the command line (bug 9829)
git-svn-id: http://svn.freepascal.org/svn/lazarus/trunk@13052 4005530d-fff6-0310-9dd1-cebe43e6787f

+++ ide/main.pp
@@ -1731,10 +1731,13 @@ begin
       AProjectFilename:=CmdLineFiles[0];
       if (CompareFileExt(AProjectFilename,'.lpr',false)=0) then
         AProjectFilename:=ChangeFileExt(AProjectFilename,'.lpi');
-      AProjectFilename:=CleanAndExpandFilename(AProjectFilename);
-      if FileExists(AProjectFilename) then begin
-        CmdLineFiles.Delete(0);
-        ProjectLoaded:=(DoOpenProjectFile(AProjectFilename,[])=mrOk);
+      // only try to load .lpi files, other files are loaded later
+      if (CompareFileExt(AProjectFilename,'.lpi',false)=0) then begin
+        AProjectFilename:=CleanAndExpandFilename(AProjectFilename);
+        if FileExists(AProjectFilename) then begin
+          CmdLineFiles.Delete(0);
+          ProjectLoaded:=(DoOpenProjectFile(AProjectFilename,[])=mrOk);
+        end;
       end;
     end;
